requirements.txt
matplotlib
Pillow
numpy